Added prototype TAP::Parser::YAMLish and associated test.
This is a stream based parser that reads from a TAP::Parser::Iterator.
It supports roughly the same YAML subset with the same semantics as
YAML::Tiny.

I think this will be renamed TAP::Parser::YAMLish::Reader and there'll
be a corresponding TAP::Parser::YAMLish::Writer which can be used by
Test::More::NextGen to generate matching YAMLish for test results.