Implemented PERL_TEST_LIBS environment variable that contains a colon separated list of library paths used to run each test script. This allows [17210] to be worked around by providing the information necessary to build -I switches for child scripts.
The other option is to manipulate PERL5LIB but that gets messy with our
current code. Particularly the logic to avoid PERL5LIB in taint mode
foils simple attempts to stuff all the includes in PERL5LIB.

I also haven't thought through all the implications of setting PERL5LIB
when PERLLIB might be in force.

I don't consider this solution final. It'd be much nicer to manipulate
PERL5LIB at least in those cases where we can do it safely.