The switches test code results in an empty string being passed into switches().  This results in the argument "" being given to Perl on VMS.  This, for whatever reason, makes Perl not work.
So eliminate passing in the empty string and at the same time simplify the code by making $args->{switches} always be an array ref.  There's only two instance of it in the code anyway.

This results in this test working on VMS except for 4 failures having to do
with differing exit codes on VMS.