- The test suite should not stop running if encountering an empty test file.
If no shebang is found, we now simply return when trying to determine if there
are extra switches to add.  The rest of the parser will now correctly report
that there was no TAP generated and subsequent tests will still be run.
Previously, the test suite halted at that point.