Rejigger the @INC propogation to be a bit simpler and remove some duplicates in our regenerated @INC.
1)  Get the differences between the current process' @INC and a fresh
process' @INC.  This captures any "use lib" or -I.  To do so we DON'T
localize PERL5LIB in _default_inc().

2)  If taint mode, encode PERL5LIB and any extra libraries as switches.
Otherwise use PERL5LIB.

Eliminate the "get the libs out of the switches" kludge.

Added a test for @INC propogation, though it doesn't catch the duplicate bug
and it's a bit of a mess.