test for new prove option --tapversion
This test adds an example TAP file that is TAP version 13
but is missing the according version directive.

By using --version=13 it is parsed as TAP13 anyways which
is verified when the contained YAML is part of generated
HTML via TAP::Formatter::HTML. This replicates a high level
use-case of prove utilizing the whole chain from cmdline
option down to TAP::Parser and formatter callbacks.