Add one more skip_all case
This test died if run with these environement variables:
- HARNESS_PERL_SWITCHES=-MDevel::Cover
- HARNESS_PERL_SWITCHES=-Mblib

This was because blib puts more directories at the beginning of @INC,
so the wibble sentinel is not the first item anymore.

Also, Devel::Cover uses blib if there is a blib directory available.

It didn't die with the this because then Devel::Cover doesn't load blib:
- HARNESS_PERL_SWITCHES=-MDevel::Cover=blib,0