Apply upstream patch:
http://perl5.git.perl.org/perl.git/commit/6359c64336d99060952232e7e300bd3c31afead8

In testargs.t in Test::Harness, don't run a world-writable file.

The test writes a file, then changes the mode, then executes it. The file needs
to be +x to be executable (on many platforms). The file will need to be +w to
be deletable on some platforms. But setting the file world writable just before
running it feels like a bad idea, given that the file's name is as predictable
as process IDs, as there's a race condition to break into the account running
perl's tests.