Link to TAP::Formatter::Color in documentation
Colored output wasn't working for me, so the first thing I looked for was documentation about what colors I should be seeing. This turned out to be in TAP::Formatter::Color, so I think it would be useful to link to it in the documentation. It also may be important for people to see that the formatter is labeled as experimental (though it's been around a while).