document that color is not default on Windows
Since colored output is not the default on Windows platforms,
this should be documented. I changed `(default)` to
`(usually default)` because there are two exceptions to the
rule of color being default, and I think it would be better
to direct users to read further documentation before assuming 
that color is always the default (and
`(default except on Windows or with non-terminal output)` was 
too long).