Allow spaces after dashes in YAMLish
When YAMLish contains an array with hashes as elements, YAML::Syck will
write the dash for the array element at a single line and the hash one
the following lines. The dash is followed by a space. This was not caught
by YAMLish reader. It is now. Also, the unit test that checks handling
spaces in YAMLish catches this behaviour now.