prove: add a --statefile=<path> option to customize the .prove file
It's inconvenient in some setups to have the .prove file be in the t/
directory, e.g. on Travis instances there's a use-case for caching the
.prove file across runs, so that you can run with
--state=failed,slow,save.

However doing so requires a Travis feature where you cache the t/
directory across runs, which is an overkill and often not what you
want, e.g. if the t/ directory will contain generated test data.

The Git project is impacted by this, currently we use a small hack to
symlink .prove to another directory:
https://github.com/git/git/commit/6272ed3194

It would be easier to just supply
--statefile=$HOME/travis-cache/prove-state, this change allows for
that.