Avoid waiting on a single test for output
When running Perl's prove with multiple jobs (-j #), the running jobs can
get stuck waiting on a single test to complete.  This occurs when a test
outputs to STDERR, but does not output to STDOUT for some time after that.
The parser iterator gets stuck waiting for can_read() on that one test,
instead of looking at all of the running tests.

Avoid this slowdown by returning, potentially early, to ensure nothing
holds up the other tests.