Generalize for EBCDIC
This had hard-coded ordinal numbers for the controls that have a
mnemonic, like \t, and hence failed on EBCDIC platforms where most of
those ordinals are different.

This commit creates a difference between EBCDIC and ASCII platforms.
On EBCDIC it initializes both the C0 and C1 controls; on ASCII, just the
C0.  My guess is that isn't a problem, but this could be easily changed
if needed.

Note that on both platforms there is a control that is an outlier.  It
is \x7f on ASCII, and \xff on EBCDIC.  This isn't initialized in either
case.  I wonder if that is a bug