stop calling import on App::Prove plugins
App::Prove plugins are passed their arguments via the load method, but
for backwards compatibility also passed the arguments with an import
call. This relies on perl ignoring the import call if the plugin does
not have such a method.

Future versions of perl are intending to throw errors for unhandled
import arguments like this. None of the plugins on CPAN include import
methods, so the easiest fix for this seems to be removing the import
call.