fix -M option to call import rather than load
The -M option to prove should load modules basically the same way perl
does. If given arguments, it should be calling the import method, not
trying to call load like it is a plugin.

Before 692c19d71fa5e18f17f2276c483f792ac7540dad, import and load would
be called both for plugins (-P) and modules (-M). This was causing
warnings and possibly in the future errors for the calls to import with
extraneous arguments. When attempting to fix this, the call to import
was entirely elimintated. That was not correct for the -M option.
Instead, -M should call import and -P should call load.